pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "jeniffer kerchner", email = "jenifferkerchner12@gmail.com"},
]
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}


[tool.pdm]
